Admin Logs

+++ classes/account.php
@@ -107,9 +107,11 @@ class Account extends Database
                 header('location: ' . HOME_URL_PREFIX . '/account?error=stmtfailed');
                 exit();
             }
-            $filePath = './assets/images/uploads/users/' . $result[0]['imagePath'];
-            if (file_exists($filePath)) {
-                unlink($filePath);
+            if ($result[0]['imagePath'] !== 'default.jpg') {
+                $filePath = './assets/images/uploads/users/' . $result[0]['imagePath'];
+                if (file_exists($filePath)) {
+                    unlink($filePath);
+                }
             }
             $stmt = $this->connect()->prepare('UPDATE user SET firstName = ?, lastName = ?, telephone = ?, gender_id = ?, state_id = ?, city = ?, zipCode = ?, imagePath = ? WHERE id = ?;');
             if (!$stmt->execute(array($data['firstName'], $data['lastName'], $data['telephone'], $data['gender'], $data['state'], $data['city'], $data['zipCode'], $finalFileName, $data['id']))) {
